+++ Samples~/MathGraph/MathGraph.cs
@@ -1,0 +1,11 @@
+using SaintsGraph;
+using UnityEngine;
+
+namespace SaintsGraphSamples
+{
+    /// <summary>A graph asset type. Everything a graph needs is to derive from NodeGraph.</summary>
+    [CreateAssetMenu(menuName = "SaintsGraph Samples/Math Graph", fileName = "MathGraph")]
+    public class MathGraph : NodeGraph
+    {
+    }
+}